Cleaned up Stu3 code difference for upsert Conflict exception
